proof-of-concept/requirements.txt
cassandra-driver==3.29.2
click==8.1.8
geomet==0.2.1.post1
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
tzdata==2024.2
